SOC meter with Arduino Nano
The circuit works in two ways; 
a) Measure ONLY the state of charge (SOC) for a cell of max 5V, 
b) Measure the SOC and enable charge or discharge the cell with two integrated relays. 
Green connectors "A6" & "A7" are placed on shunt resistor, red connector on + of the battery cell, black connector on GND. When on mode (b), black connector is placed on GND charger and GND load, the yellow "load" connector on + of the load and the yellow "line in" on the exit of the shunt resistor (with the green "A7" connector). There is a data cable/connector that sends via <wire.h> the integer SOC to the other Arduino Mega, where can be displayed graphical with the Nextion display. 
When first time plugged in (USB), the TFT display on the device after 5 seconds welcome menu waits for the min and max voltage. This values are changed with two potentiometers, set by the black button next to them. The green LED shows that the charging relay is connected. The parameters are stored in the Nano eeprom, the data can be erased by holding (more than 5 sec) the red button. The two way switch is optional and not in use for now.